Constructor for R6 ctsmrTMB class

Call ctsmrTMB$new() to construct a new instance of the class object for specifying a new stochastic state space model.

Value

A model object of class 'ctsmrTMB'

Methods

Public methods

ctsmrTMB$new()

ctsmrTMB$add_systems()

ctsmrTMB$add_observations()

ctsmrTMB$add_observation_variances()

ctsmrTMB$add_inputs()

ctsmrTMB$add_parameters()

ctsmrTMB$add_algebraics()

ctsmrTMB$add_constants()

ctsmrTMB$set_initial_state()

ctsmrTMB$set_lamperti()

ctsmrTMB$set_modelname()

ctsmrTMB$set_cppfile_directory()

ctsmrTMB$set_map()

ctsmrTMB$get_systems()

ctsmrTMB$get_observations()

ctsmrTMB$get_observation_variances()

ctsmrTMB$get_algebraics()

ctsmrTMB$get_parameters()

ctsmrTMB$construct_nll()

ctsmrTMB$estimate()

ctsmrTMB$predict()

ctsmrTMB$print()

ctsmrTMB$summary()

ctsmrTMB$plot()

ctsmrTMB$clone()

Method new()

Construct a new ctsmrTMB model object.

Usage

ctsmrTMB$new()

Method add_systems()

Define and add multiple stochastic differential equation governing the process of individual state variables on the form

d<state> ~ f(t,<states>,<inputs>) * dt + g1(t,<states>,<inputs>) * dw1 + g2(t,<states>,<inputs>) * dw2 + ...

where f is the drift, and g1, g2, ... are diffusions, with differential brownian motions dw1, dw2, ...

Usage

ctsmrTMB$add_systems(form, ...)

Arguments

form

formula specifying the stochastic differential equation to be added to the system.

...

additional formulas similar to form for specifying multiple equations at once.

Examples

# Specify Ornstein-Uhlenbeck Process
add_systems(dX ~ theta * (mu - X + u) * dt + sigma * dw)

# Specify Lokta-Volterra System of Equations
add_systems( dN ~ ( r * N * (1-N/K) - c*N*P/(N+Nbar) ) * dt + sigmaN * N * dw1,
             dP ~ ( eps*c*N*P/(N+Nbar) - mu * P ) * dt + sigmaP * P * dw2 )

Method add_observations()

Define and add a relationship between an observed variable and system states. The observation equation takes the form

<observation> ~ h(t,<states>,<inputs>) + e)

where h is the observation function, and e is normally distributed noise with zero mean and variance to be specified. The observation variable should be present in the data provided when calling estimate(.data) for parameter estimation.

Usage

ctsmrTMB$add_observations(form, ..., obsnames = NULL)

Arguments

form

formula class specifying the obsevation equation to be added to the system.

...

additional formulas identical to form to specify multiple observation equations at a time.

obsnames

character vector specifying the name of the observation. When the observation left-hand side consists of more than just a single variable name (when its class is 'call' instead of 'name') it will be given a name on the form obs__# where # is a number, unless obsnames is provided.

Examples

#Specify observation directly as a latent state
add_observations(y ~ x)

Specify observation as the sum of exponentials of two latent states
add_observations(y ~ exp(x1) + exp(x2))

Method add_observation_variances()

Specify the variance of an observation equation.

A defined observation variable y in e.g. add_observations(y ~ h(t,<states>,<inputs>) is pertubed by Gaussian noise with zero mean and variance to-be specified using add_observation_variances(y ~ p(t,<states>,<inputs>). We can for instance declare add_observation_variances(y ~ sigma_x^2 where sigma_x is a fixed effect parameter to be declared through add_parameters.

Usage

ctsmrTMB$add_observation_variances(form, ...)

Arguments

form

formula class specifying the obsevation equation to be added to the system.

...

additional formulas identical to form to specify multiple observation equations at a time.

Method add_inputs()

Declare variables as data inputs

Declare whether a variable contained in system, observation or observation variance equations is an input variable. If e.g. the system equation contains an input variable u then it is declared using add_inputs(u). The input u must be contained in the data.frame .data provided when calling the estimate or predict methods.

Usage

ctsmrTMB$add_inputs(...)

Arguments

...

unquoted variable names that specifies the name of input variables in the defined system.

Method add_parameters()

Declare variables as fixed effects and specify their initial value, lower and upper bound used when calling the maximum likelihood optimization.

There are two ways to declare parameters. You can declare parameters using formulas i.e. add_parameters( theta ~ c(1,0,10), mu ~ c(0,-10,10) ), where the values are initial, lower and upper bound respectively. Alternatively you can provide a 3-column matrix where rows corresponds to different parameters, and the parameter names are provided as rownames of the matrix.

Usage

ctsmrTMB$add_parameters(...)

Arguments

...

formula whose left-hand side is the parameter name, and right hand side is a vector of length 3 with inital value, lower and upper bound respectively. You can provide multiple parameters at once by seperating formulas with comma.

parameter.matrix

matrix of 3 columns where rows correspond to variables. The variable names must be provided as rownames to the matrix. The columns are initial value, lower and upper bound respectively.

Method add_algebraics()

Add algebraic relations.

Algebraic relations is a convenient way to transform parameters in your equations. In the Ornstein-Uhlenbeck process the rate parameter theta is always positive, so estimation in the log-domain is a good idea. Instead of writing exp(theta) directly in the system equation one can transform into the log domain using the algebraic relation add_algebraics(theta ~ exp(logtheta)). All instances of theta is replaced by exp(logtheta) when compiling the C++ function. Note that you must provide values for logtheta now instead of theta when declaring parameters through add_parameters

Usage

ctsmrTMB$add_algebraics(form, ...)

Arguments

form

formula specifying the stochastic differential equation(s) to be added to the system.

...

additional formulas similar to form for specifying multiple equations at once.

Method add_constants()

Declare variables as scalar constants.

Usage

ctsmrTMB$add_constants(...)

Arguments

...

additional formulas similar to form for specifying multiple constants at once.

form

formula whose left-hand side is the constant variable name, and the right-hand side is its numeric value.

Method set_initial_state()

Declare the initial state values i.e. mean and covariance for the system states.

Usage

ctsmrTMB$set_initial_state(mean, cov, estimate = FALSE)

Arguments

mean

numeric vector of size equal to n (number of system states)

cov

matrix (symmetric positive semi-definite) of dimensions n^2.

estimate

boolean value which indicates whether or not the initial conditions shall be estimated as fixed effects parameters. The provided mean and covariance are then used as initial guesses

Method set_lamperti()

Set a Lamperti Transformation

If the provided system equations have state dependent diffusion in of a few available ways then it is advantageous to perform a transformation to remove the state dependence. This comes at the cost of a more complicated drift function. The following types of state-dependence is currently supported

If no transformation is desired choose 'identity' (default).

If the diffusion term is proportional to x * dw, then a log-transform is available

If the diffusion term is proportional to x * (1-x) * dw, then a logit-transform is available

If the diffusion term is proportional to sqrt(x * (1-x)) * dw, then a sqrt-logit-transform is available

If the diffusion term is proportional to x * (1-x^a) * dw, for a>0 then a power-logit transform is available

Usage

ctsmrTMB$set_lamperti(transform, states = NULL)

Arguments

transform

character vector - one of either "identity, "log", "logit", "sqrt-logit", or "power-logit"

states

a vector of the state names for which the chosen transformation is desired. The default (NULL) is to apply the transformation to all state equations.

Method set_modelname()

Set modelname used to create the C++ file for TMB

When calling TMB::MakeADFun the (negative log) likelihood function is created in the directory specified by the set_cppfile_directory method with name <modelname>.cpp

Usage

ctsmrTMB$set_modelname(name)

Arguments

name

string defining the model name.

Method set_cppfile_directory()

Set the path directory where the constructed C++ file is created. You should specify the entire path, unless you want to construct a subfolder in the current working directory - then you can call e.g. set_cppfile_directory("folder_in_current_wd").

Usage

ctsmrTMB$set_cppfile_directory(directory)

Arguments

directory

string specifying a local directory

Method set_map()

Enable maximum a posterior (MAP) estimation.

Adds a maximum a posterior contribution to the (negative log) likelihood function by evaluating the fixed effects parameters in a multivariate Gaussian with mean and covariance as provided.

Usage

ctsmrTMB$set_map(mean, cov)

Arguments

mean

mean vector of the Gaussian prior parameter distribution

cov

covariance matrix of the Gaussian prior parameter distribution

Method get_systems()

Retrieve system equations.

Usage

ctsmrTMB$get_systems()

Method get_observations()

Retrieve observation equations.

Usage

ctsmrTMB$get_observations()

Method get_observation_variances()

Retrieve observation variances

Usage

ctsmrTMB$get_observation_variances()

Method get_algebraics()

Retrieve algebraic relations

Usage

ctsmrTMB$get_algebraics()

Method get_parameters()

Get initial (and estimated) parameters.

Usage

ctsmrTMB$get_parameters()

Method construct_nll()

Construct and extract function handlers for the negative log likelihood function.

The handlers from TMB's MakeADFun are constructed and returned. This enables the user to e.g. determine their own optimization algorithm to use for estimation.

Usage

ctsmrTMB$construct_nll(
  data,
  method = "ekf",
  ode.solver = "euler",
  ode.timestep = min(diff(data$t)),
  loss = "quadratic",
  loss_c = 3,
  compile = FALSE
)

Arguments

data

data.frame containing time-vector 't', observations and inputs. The observations can take NA-values.

method

character vector - one of either "ekf", "ukf" or "tmb". Sets the estimation method. The package has three available methods implemented:

The natural TMB-style formulation where latent states are considered random effects and are integrated out using the Laplace approximation. This method only yields the gradient of the (negative log) likelihood function with respect to the fixed effects for optimization. The method is slower although probably has some precision advantages, and allows for non-Gaussian observation noise (not yet implemented). One-step / K-step residuals are not yet available in the package.

(Continous-Discrete) Extended Kalman Filter where the system dynamics are linearized to handle potential non-linearities. This is computationally the fastest method.

(Continous-Discrete) Unscented Kalman Filter. This is a higher order non-linear Kalman Filter which improves the mean and covariance estimates when the system display high nonlinearity, and circumvents the necessity to compute the jacobian of the drift and observation functions.

All package features are currently available for the kalman filters, while TMB is limited to parameter estimation. In particular, it is straight-forward to obtain k-step-ahead predictions with these methods (use the predict S3 method), and stochastic simulation is also available in the cases where long prediction horizons are sought, where the normality assumption will be inaccurate.

ode.solver

Sets the ODE solver used in the Kalman Filter methods for solving the moment differential equations. The default "euler" is the Forward Euler method, alternatively the classical 4th order Runge Kutta method is available via "rk4".

ode.timestep

numeric value. Sets the time step-size in numerical filtering schemes. The defined step-size is used to calculate the number of steps between observation time-points as defined by the provided data. If the calculated number of steps is larger than N.01 where N is an integer, then the time-step is reduced such that exactly N+1 steps is taken between observations The step-size is used in the two following ways depending on the chosen method:

Kalman filters: The time-step is used as the step-size in the numerical Forward-Euler scheme to compute the prior state mean and covariance estimate as the final time solution to the first and second order moment differential equations.

TMB method: The time-step is used as the step-size in the Euler-Maruyama scheme for simulating a sample path of the stochastic differential equation, which serves to link together the latent (random effects) states.

loss

character vector. Sets the loss function type (only implemented for the kalman filter methods). The loss function is per default quadratic in the one-step residauls as is natural when the Gaussian (negative log) likelihood is evaluated, but if the tails of the distribution is considered too small i.e. outliers are weighted too much, then one can choose loss functions that accounts for this. The three available types available:

Quadratic loss (quadratic).

Quadratic-Linear (huber)

Quadratic-Constant (tukey)

The cutoff for the Huber and Tukey loss functions are determined from a provided cutoff parameter loss_c. The implementations of these losses are approximations (pseudo-huber and sigmoid approxmation respectively) for smooth derivatives.

loss_c

cutoff value for huber and tukey loss functions. Defaults to c=3

compile

boolean value. The default (FALSE) is to not compile the C++ objective function but assume it is already compiled and corresponds to the specified model object. It is the user's responsibility to ensure correspondence between the specified model and the precompiled C++ object. If a precompiled C++ object is not found in the specified directory i.e. in <cppfile_directory>/<modelname>/(dll/so) then the compile flag is set to TRUE. If the user makes changes to system equations, observation equations, observation variances, algebraic relations or lamperi transformations then the C++ object should be recompiled.

Method estimate()

Estimate the fixed effects parameters in the specified model.

Usage

ctsmrTMB$estimate(
  data,
  method = "ekf",
  ode.solver = "euler",
  ode.timestep = min(diff(data$t)),
  loss = "quadratic",
  loss_c = 3,
  use.hessian = FALSE,
  compile = FALSE,
  control = list(trace = 1)
)

Arguments

data

data.frame containing time-vector 't', observations and inputs. The observations can take NA-values.

method

character vector - one of either "ekf", "ukf" or "tmb". Sets the estimation method. The package has three available methods implemented:

The natural TMB-style formulation where latent states are considered random effects and are integrated out using the Laplace approximation. This method only yields the gradient of the (negative log) likelihood function with respect to the fixed effects for optimization. The method is slower although probably has some precision advantages, and allows for non-Gaussian observation noise (not yet implemented). One-step / K-step residuals are not yet available in the package.

(Continous-Discrete) Extended Kalman Filter where the system dynamics are linearized to handle potential non-linearities. This is computationally the fastest method.

(Continous-Discrete) Unscented Kalman Filter. This is a higher order non-linear Kalman Filter which improves the mean and covariance estimates when the system display high nonlinearity, and circumvents the necessity to compute the jacobian of the drift and observation functions.

All package features are currently available for the kalman filters, while TMB is limited to parameter estimation. In particular, it is straight-forward to obtain k-step-ahead predictions with these methods (use the predict S3 method), and stochastic simulation is also available in the cases where long prediction horizons are sought, where the normality assumption will be inaccurate.

ode.solver

Sets the ODE solver used in the Kalman Filter methods for solving the moment differential equations. The default "euler" is the Forward Euler method, alternatively the classical 4th order Runge Kutta method is available via "rk4".

ode.timestep

numeric value. Sets the time step-size in numerical filtering schemes. The defined step-size is used to calculate the number of steps between observation time-points as defined by the provided data. If the calculated number of steps is larger than N.01 where N is an integer, then the time-step is reduced such that exactly N+1 steps is taken between observations The step-size is used in the two following ways depending on the chosen method:

Kalman filters: The time-step is used as the step-size in the numerical Forward-Euler scheme to compute the prior state mean and covariance estimate as the final time solution to the first and second order moment differential equations.

TMB method: The time-step is used as the step-size in the Euler-Maruyama scheme for simulating a sample path of the stochastic differential equation, which serves to link together the latent (random effects) states.

loss

character vector. Sets the loss function type (only implemented for the kalman filter methods). The loss function is per default quadratic in the one-step residauls as is natural when the Gaussian (negative log) likelihood is evaluated, but if the tails of the distribution is considered too small i.e. outliers are weighted too much, then one can choose loss functions that accounts for this. The three available types available:

Quadratic loss (quadratic).

Quadratic-Linear (huber)

Quadratic-Constant (tukey)

The cutoff for the Huber and Tukey loss functions are determined from a provided cutoff parameter loss_c. The implementations of these losses are approximations (pseudo-huber and sigmoid approxmation respectively) for smooth derivatives.

loss_c

cutoff value for huber and tukey loss functions. Defaults to c=3

use.hessian

boolean value. The default (TRUE) causes the optimization algorithm stats::nlminb to use the fixed effects hessian of the (negative log) likelihood when performing the optimization. This feature is only available for the kalman filter methods without any random effects.

compile

boolean value. The default (FALSE) is to not compile the C++ objective function but assume it is already compiled and corresponds to the specified model object. It is the user's responsibility to ensure correspondence between the specified model and the precompiled C++ object. If a precompiled C++ object is not found in the specified directory i.e. in <cppfile_directory>/<modelname>/(dll/so) then the compile flag is set to TRUE. If the user makes changes to system equations, observation equations, observation variances, algebraic relations or lamperi transformations then the C++ object should be recompiled.

control

list of control parameters parsed to nlminb as its control argument. See ?stats::nlminb for more information

Method predict()

Perform prediction/filtration to obtain state mean and covariance estimates. The predictions are obtained by solving the moment equations n.ahead steps forward in time when using the current step posterior state estimate as the initial condition.

Usage

ctsmrTMB$predict(
  data = NULL,
  pars = NULL,
  k.ahead = 1,
  return.k.ahead = NULL,
  method = "ekf",
  ode.timestep = min(diff(data$t)),
  ode.solver = "euler",
  x0 = NULL,
  p0 = NULL,
  return.covariance = TRUE,
  compile = FALSE
)

Arguments

data

data.frame containing time-vector 't', observations and inputs. The observations can take NA-values.

pars

fixed parameter vector parsed to the objective function for prediction/filtration. The default parameter values used are the initial parameters provided through add_parameters, unless the estimate function has been run, then the default values will be those at the found optimum.

k.ahead

integer specifying the desired number of time-steps (as determined by the provided data time-vector) for which predictions are made (integrating the moment ODEs forward in time without data updates).

return.k.ahead

numeric vector of integers specifying which n.ahead predictions to that should be returned.

method

character vector - one of either "ekf", "ukf" or "tmb". Sets the estimation method. The package has three available methods implemented:

ONLY EKF IS IMPLEMENTED CURRENTLY!!

The natural TMB-style formulation where latent states are considered random effects and are integrated out using the Laplace approximation. This method only yields the gradient of the (negative log) likelihood function with respect to the fixed effects for optimization. The method is slower although probably has some precision advantages, and allows for non-Gaussian observation noise (not yet implemented). One-step / K-step residuals are not yet available in the package.

(Continous-Discrete) Extended Kalman Filter where the system dynamics are linearized to handle potential non-linearities. This is computationally the fastest method.

(Continous-Discrete) Unscented Kalman Filter. This is a higher order non-linear Kalman Filter which improves the mean and covariance estimates when the system display high nonlinearity, and circumvents the necessity to compute the jacobian of the drift and observation functions.

All package features are currently available for the kalman filters, while TMB is limited to parameter estimation. In particular, it is straight-forward to obtain k-step-ahead predictions with these methods (use the predict S3 method), and stochastic simulation is also available in the cases where long prediction horizons are sought, where the normality assumption will be inaccurate.

ode.timestep

numeric value. Sets the time step-size in numerical filtering schemes. The defined step-size is used to calculate the number of steps between observation time-points as defined by the provided data. If the calculated number of steps is larger than N.01 where N is an integer, then the time-step is reduced such that exactly N+1 steps is taken between observations The step-size is used in the two following ways depending on the chosen method:

Kalman filters: The time-step is used as the step-size in the numerical Forward-Euler scheme to compute the prior state mean and covariance estimate as the final time solution to the first and second order moment differential equations.

TMB method: The time-step is used as the step-size in the Euler-Maruyama scheme for simulating a sample path of the stochastic differential equation, which serves to link together the latent (random effects) states.

ode.solver

Sets the ODE solver used in the Kalman Filter methods for solving the moment differential equations. The default "euler" is the Forward Euler method, alternatively the classical 4th order Runge Kutta method is available via "rk4".

x0

numeric vector containing the initial expected state values

p0

numeric matrix contianing the initial state covariances

return.covariance

booelan value to indicate whether the covariance (instead of the correlation) should be returned.

compile

boolean value. The default (FALSE) is to not compile the C++ objective function but assume it is already compiled and corresponds to the specified model object. It is the user's responsibility to ensure correspondence between the specified model and the precompiled C++ object. If a precompiled C++ object is not found in the specified directory i.e. in <cppfile_directory>/<modelname>/(dll/so) then the compile flag is set to TRUE. If the user makes changes to system equations, observation equations, observation variances, algebraic relations or lamperi transformations then the C++ object should be recompiled.

Returns

A data.frame that contains for each time step the posterior state estimate at that time.step (k = 0), and the prior state predictions (k = 1,...,n.ahead). If return.covariance = TRUE then the state covariance/correlation matrix is returned, otherwise only the marginal variances are returned.

Method print()

Function to print the model object

Usage

ctsmrTMB$print()

Method summary()

Summary function for fit

Usage

ctsmrTMB$summary(correlation = FALSE)

Arguments

correlation

boolean value. The default (FALSE) is to not provide the fixed effects parameter correlation matrix.

Method plot()

Function to print the model object

Usage

ctsmrTMB$plot(
  plot.obs = 1,
  use.ggplot = FALSE,
  extended = FALSE,
  ggtheme = getggplot2theme()
)

Arguments

plot.obs

integer value. The value determines which state residual plot is shown, when the states are numbered by the order in which they were defined in the model object.

use.ggplot

boolean value. The default FALSE is to use base plots in R, if TRUE then ggplot2 is used to show and return a list of all residual state plots.

extended

boolean value. No functionality yet.

ggtheme

ggplot theme. This determines the theme used to construct the ggplot2 if use.ggplot=TRUE.

Method clone()

The objects of this class are cloneable with this method.

Usage

ctsmrTMB$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples