Predict state estimates on ctsmrTMB.fit object.

Predict state estimates on ctsmrTMB.fit object.

# S3 method for ctsmrTMB.fit
predict(
  fit,
  data = NULL,
  x0 = NULL,
  p0 = NULL,
  n.step.ahead = 1,
  ode.timestep = NULL,
  return.state.dispersion = FALSE,
  covariance = FALSE,
  give.only.n.step.ahead = FALSE,
  use.simulation = FALSE,
  sim.timestep = NULL,
  returned.sim.dt = NULL,
  n.sim = 100,
  quantiles = c(0.05, 0.5, 0.95),
  return.full.simulations = FALSE
)

Arguments

fit

An ctsmrTMB.fit object from calling estimate method on ctsmrTMB object.

data

data.frame used to predict and update the model on which fit is based.

use.simulation

Boolean. If TRUE simulates the stochastic model and returns quantiles. if FALSE uses and returns the solution values for the first and second order moment differential equations. Default: FALSE

n.sim

Integer. Number of simulations performed if use.simulation is TRUE. Default: 100

quantiles

The returned state distribution quantiles if use.simulation is TRUE.

n.step

Integer. Number of prediction steps before updating. Default: 1

Value

Prediction from model